[FEAT](listings): Implementacion de busqueda dinamica, paginacion y visibilidad de hidden por rol

+++ PTruequeU/DTOs/Listings/ListingSearchDto.cs
@@ -6,10 +6,12 @@ namespace PTruequeU.DTOs.Listings
     {
         public string? Keyword { get; set; }
         public Guid? CategoryId { get; set; }
-        public double? MinPrice { get; set; }
-        public double? MaxPrice { get; set; }
+        public decimal? MinPrice { get; set; }
+        public decimal? MaxPrice { get; set; }
+
         public ListingCondition? Condition { get; set; }
         public ListingState? State { get; set; }
+
         public DateTime? PostedAfter { get; set; }
         public DateTime? PostedBefore { get; set; }
